move files to appropriate folders
